refactor: move widget tests to nx shared
