+++ bciers/libs/shared/components/src/lib/form/widgets/FileWidget.test.tsx
@@ -2,8 +2,8 @@ import { userEvent } from "@testing-library/user-event";
 import { render, screen, waitFor } from "@testing-library/react";
 import { RJSFSchema } from "@rjsf/utils";
 import FormBase from "@components/form/FormBase";
-import { actionHandler, useSession } from "@/tests/setup/mocks";
-import { Session } from "@/tests/setup/types";
+import { actionHandler, useSession } from "@tests/mocks";
+import { Session } from "@tests/types";
 import { checkNoValidationErrorIsTriggered } from "@/tests/helpers/form";
 
 const fileFieldLabel = "FileWidget test field";
